refactor(predict): decompose PredictMarketDetails into focused sub-components
Extract 8 sub-components and 3 custom hooks from the monolithic
PredictMarketDetails view (1,391 → 492 lines) to improve maintainability
and readability.

Components: PredictMarketDetailsAbout, Actions, Header, Outcomes,
Positions, Status, TabBar, TabContent
Hooks: useChartData, useOpenOutcomes, useOutcomeResolution

All existing tests (107 cases) pass without modification.
