+++ app/components/UI/Predict/views/PredictMarketDetails/components/PredictMarketDetailsAbout/index.ts
@@ -1,0 +1,2 @@
+export { default } from './PredictMarketDetailsAbout';
+export type { PredictMarketDetailsAboutProps } from './PredictMarketDetailsAbout';
