+++ app/components/UI/Predict/views/PredictMarketDetails/components/PredictMarketDetailsActions/index.ts
@@ -1,0 +1,2 @@
+export { default } from './PredictMarketDetailsActions';
+export type { PredictMarketDetailsActionsProps } from './PredictMarketDetailsActions';
