+++ app/components/UI/Predict/views/PredictMarketDetails/components/PredictMarketDetailsHeader/index.ts
@@ -1,0 +1,2 @@
+export { default } from './PredictMarketDetailsHeader';
+export type { PredictMarketDetailsHeaderProps } from './PredictMarketDetailsHeader';
