+++ app/components/UI/Predict/views/PredictMarketDetails/components/PredictMarketDetailsOutcomes/index.ts
@@ -1,0 +1,2 @@
+export { default } from './PredictMarketDetailsOutcomes';
+export type { PredictMarketDetailsOutcomesProps } from './PredictMarketDetailsOutcomes';
