+++ app/components/UI/Predict/views/PredictMarketDetails/components/PredictMarketDetailsPositions/index.ts
@@ -1,0 +1,2 @@
+export { default } from './PredictMarketDetailsPositions';
+export type { PredictMarketDetailsPositionsProps } from './PredictMarketDetailsPositions';
