+++ app/components/UI/Predict/views/PredictMarketDetails/components/PredictMarketDetailsStatus/index.ts
@@ -1,0 +1,2 @@
+export { default } from './PredictMarketDetailsStatus';
+export type { PredictMarketDetailsStatusProps } from './PredictMarketDetailsStatus';
