+++ app/components/UI/Predict/views/PredictMarketDetails/components/PredictMarketDetailsTabBar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './PredictMarketDetailsTabBar';
+export type { PredictMarketDetailsTabBarProps } from './PredictMarketDetailsTabBar';
